+++ packages/skeleton-react/src/lib/components/Progress/Progress.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import * as progress from '@zag-js/progress';
 import { normalizeProps, useMachine } from '@zag-js/react';
 import { useId } from 'react';