+++ packages/typelens/test/fixtures/multiple.ts
@@ -1,0 +1,9 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+interface InterfaceOne {
+	property: string;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+interface InterfaceTwo {
+	property: string;
+}